v0.2
GUIDBackup
Make a scheduled task with this file in the program entry, set it to run with highest privledge and run when user is not logged in. It will do a backup of allcritical drives to an external drive that has a label containing the string "Backup".

So "Backup A", "Backup 1" are all valid names. The backup job itself attaches the disk to drive letter K on briefly and then disconnects the drive letter to perform the backup with the guid only.